requirements.txt
asgiref==3.10.0
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.3
tzdata==2025.2
pillow==11.0.0
drf-spectacular==0.27.1
django-cleanup==9.0.0
celery==5.4.0
redis==5.0.8
celery==5.4.0
redis==5.0.8
django-celery-beat>=2.7.0
python-json-logger==2.0.7
requests